Restructure Stanza classes
