+++ app/models/ocelot/Stanza.scala
@@ -16,33 +16,6 @@
 
 package models.ocelot
 
-import play.api.libs.json._
-import play.api.libs.functional.syntax._
+trait Stanza
 
-sealed trait Stanza
-
-
-/**
- * Define Callout Stanza
- */
-case class CalloutStanza(
-                        noteType: String,
-                        text: Int,
-                        next: Seq[String],
-                        stack: Boolean
-                        ) extends Stanza
-
-object CalloutStanza {
-
-  implicit val calloutReads: Reads[CalloutStanza] = {
-
-    (( JsPath \ "noteType" ).read[String] and
-      ( JsPath \ "text" ).read[Int] and
-      ( JsPath \ "next" ).read[Seq[String]] and
-      ( JsPath \ "stack" ).read[Boolean]
-      )(CalloutStanza.apply _)
-
-  }
-
-}
